+++ app/(afterLogin)/post/detail/[postId]/_components/section/commentSection.tsx
@@ -1,11 +1,40 @@
+"us"
+
 import LikeToComment from "../comment/_components/likeToComment"
 import CommentBox from "../comment/commentBox"
 import CreateComment from "../comment/createComment"
 import { CommentSectionProps } from "./type"
+import { useQuery } from "@tanstack/react-query"
+import { deleteComment, getComment } from "@/apis/commentApi"
+import { useModalStore } from "@/libs/zustand/modal"
+import DeleteModal from "@/components/commons/deleteModal"
+import { useCommentIdStore } from "@/libs/zustand/comment"
+import { Comment } from "../comment/type"
 
-const CommentSection = ({ postId, post, comments, refetchComments }: CommentSectionProps) => {
+const CommentSection = ({ postId, post }: CommentSectionProps) => {
+    const { isDelete, setIsDelete } = useModalStore()
+    const { saveCommentId } = useCommentIdStore()
+
+    const { data: comments = [], refetch: refetchComments } = useQuery<Comment[], Error>({
+        queryKey: ["comments", { postId: Number(postId), page: 0, size: 9999 }],
+        queryFn: () => getComment({ postId: Number(postId), page: 0, size: 9999 }),
+    })
+
+    const handleDelete = async (commentId: number) => {
+        setIsDelete(false)
+        await deleteComment({ commentId: commentId })
+        refetchComments()
+    }
+
     return (
         <>
+            <DeleteModal
+                isOpen={isDelete}
+                onClick={() => handleDelete(saveCommentId)}
+                onLeftClick={() => setIsDelete(false)}
+                title="댓글"
+                context="댓글"
+            />
             <LikeToComment likes={post.likeCount} comments={comments.length} />
             <CreateComment postId={postId} refetch={refetchComments} />
             <div className="flex items-center justify-center">
